requirements.txt
asgiref
coverage
Django
djangorestframework
pytz
sqlparse
tzdata
python-dotenv